Still non-functional. Some refactoring of ct_bact experiment in task.js.
The javascript for the ct_bact experiment is coming into line with the template,
and I've been making simplifying and clarifying changes. The original code is still
sitting at the bottom of the file.

Also, a cutting up of task.js into a few components would probably be desireable.

Next steps:
- get consent form in it's proper place
- clean up js in introductions
- remove old goTo() buttons from various pages
- debug main task
- delete original code (which is still sitting there for ref.)
- integrate questionaire forms & their js